Seperate Transition class into its own file
